refactor(trusty-search): split service_unit.rs test bodies into sibling files
service_unit.rs was 501 SLOC, one over the 500 production cap
(scripts/check_line_cap.sh), blocking PR #4919. Moves the three inline
`#[cfg(test)]` modules (parse_truthy_bool tests, the macOS launchd-unit
tests, and the cli_tests clap-wiring tests) into sibling
service_unit_*_tests.rs files, wired back in via `#[path]` — the same
split convention already used for service.rs/service_tests.rs in this
directory. Filenames ending `_tests.rs` classify as test files (3000-SLOC
cap) under scripts/check_line_cap.sh, so this is a real split, not an
allowlist exemption.

Each stub mod keeps its original nesting depth and cfg gates, so
`use super::*;` inside the moved files resolves identically to before.
No public API changed; no caller edits needed.
